refactor(form): rename model file
