+++ src/components/Form/readme.md
@@ -3,11 +3,11 @@
 Todos:
 
 - [x] Support validation
-  - [ ] Cleanup nested form promise handler
 - [ ] Support Custom Validation composition
 - [x] Able to point out field that failed
 - [x] Nesting forms
-- [ ] Loose coupling of form logics and UI renderings
+- [ ] Form Component with renderProps containing isError, isTouched down to field level
+- [x] Loose coupling of form logics and UI renderings
 - [x] Reactive upon value change
 - [x] Support initial value
 - [x] Support reset
